Research on B Cell Algorithm for Learning to Rank Method Based on Parallel Strategy
In information retrieval, users need to find documents with high
relevance from a huge number of documents on the basis of the input
query. How to rank the documents according to relevance in the system
and meet user needs become the import issues to be solved in information
retrieval. The methods called learning to rank using machine learning
techniques to generate ranking function automatically are applied
extensively. A parallel B cell algorithm is proposed using clonal
selection mechanism in biological immunity for reference and applied to
learning to rank, thus an optimal ranking function can be determined.
The novel algorithm is compared with traditional ones through
experiment. The experimental result proves that the proposed algorithm
outperforms the traditional ones in respect of accuracy, learning time
and convergence rate and is able to determine an optimal ranking
function, it is an effective learning to rank algorithm.